LOTS of refactoring going on here...
